Reorganize JavaFX components in folder
